requirements.txt
pyhiveapi
googleapiclient
google-api-python-client
google-auth-httplib2
google-auth-oauthlib